requirements.txt
# Gmail API
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.122.0

# Google Cloud Text-to-Speech
google-cloud-texttospeech==2.16.3

# Google Gemini AI
google-generativeai==0.4.1

# Email handling
email-validator==2.1.1

# Utilities
python-dotenv==1.0.1
